requirements.txt
boto3>=1.34.0
botocore>=1.34.0
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
sqlparse>=0.4.4
pyyaml>=6.0.0
jsonschema>=4.0.0
